requirements.txt
absl-py==0.2.2
astor==0.6.2
audioread==2.1.6
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
bokeh==0.12.16
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.12.0
html5lib==0.9999999
intervaltree==2.1.0
ipython==5.7.0
ipython-genutils==0.2.0
Jinja2==2.10
joblib==0.11
kiwisolver==1.0.1
librosa==0.6.1
llvmlite==0.23.2
magenta-gpu==0.3.8
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mido==1.2.6
mir-eval==0.4
mock==2.0.0
numba==0.38.1
numpy==1.14.3
packaging==17.1
pandas==0.23.0
pathlib2==2.3.2
pbr==4.0.3
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pkg-resources==0.0.0
pretty-midi==0.2.8
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-rtmidi==1.1.0.dev0
pytz==2018.4
PyYAML==3.12
resampy==0.2.0
rtmidi==2.3.3
scandir==1.7
scikit-learn==0.19.1
scipy==1.1.0
selenium==3.141.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sortedcontainers==2.0.3
subprocess32==3.5.1
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
tornado==5.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.14.1
